eshu ready for parallel scrapping

+++ eshu/cdn.bash
@@ -9,7 +9,7 @@ set -o errexit
 set -o nounset
 
 #(a.k.a set -x) to trace what gets executed
-set -o xtrace
+#set -o xtrace
 
 # in scripts to catch mysqldump fails 
 set -o pipefail
@@ -29,7 +29,7 @@ pidfile=$HOME/.${__base}.pid
 if [ -f ${pidfile} ]; then
    #verify if the process is actually still running under this pid
    oldpid=`cat ${pidfile}`
-   result=`ps -ef | grep ${oldpid} | grep ${scriptname}`  
+   result=`ps -ef | grep ${oldpid} | grep ${__base}  || true`
 
    if [ -n "${result}" ]; then
      echo "Script already running! Exiting"
@@ -38,7 +38,7 @@ if [ -f ${pidfile} ]; then
 fi
 
 #grab pid of this process and update the pid file with it
-pid=`ps -ef | grep ${__base} | head -n1 |  awk ' {print $2;} '`
+pid=`ps -ef | grep ${__base} | grep -v 'vi ' |  head -n1 |  awk ' {print $2;} '`
 echo ${pid} > ${pidfile}
 
 
@@ -62,34 +62,72 @@ echo Begin `date`  .....
 ### BEGIN SCRIPT ###############################################################
 
 electionID=`echo $__base | cut -d '.' -f 1`
+pauseSecs=2
+restSecs=10
+continueFlag=/tmp/$__base.continue
+stopFlag=/tmp/$__base.stop
 
-for lang in e f; do 
-	dataDir="$__dir/$electionID/$lang/"
-	workDir="$dataDir/work/"
-	readyDir="$dataDir/ready"
+touch $continueFlag
 
-	mkdir -p $workDir $readyDir
+if [[ -f $stopFlag ]]; then
+	echo Stop Flag active.  To remove
+	echo rm $stopFlag
+	exit 1
+fi
 
-	sourceUrl="http://enr.elections.ca/ElectoralDistricts.aspx?lang=$lang"
-	sourceFile="$dataDir/source.html"
-	curl $sourceUrl > $sourceFile
+while [[ -f $continueFlag && ! -f $stopFlag ]]; do 
+	for lang in e f; do 
+		dataDir="$__dir/data/$electionID/$lang/"
+		workDir="$dataDir/work/"
+		readyDir="$dataDir/ready"
+	
+		mkdir -p $workDir $readyDir
+	
+		sourceUrl="http://enr.elections.ca/ElectoralDistricts.aspx?lang=$lang"
+		sourceFile="$dataDir/source.html"
 
-	ridingList="$dataDir/ridingIDsList.txt"
-	grep '<li><a href="ElectoralDistricts.aspx?ed=' $sourceFile | cut -d '=' -f 3 | cut -d '&' -f 1 > $ridingList
+		echo; echo Getting riding list...
+		curl $sourceUrl > $sourceFile
+		echo; echo Done. ; echo
+	
+		ridingList="$dataDir/ridingIDsList.txt"
+		grep '<li><a href="ElectoralDistricts.aspx?ed=' $sourceFile | cut -d '=' -f 3 | cut -d '&' -f 1 > $ridingList
+	
+		for identifier in `cat $ridingList`; do
+			ridingUrl="http://enr.elections.ca/ElectoralDistricts.aspx?ed=$identifier&lang=$lang"
+			ridingFile=$workDir/$identifier.html
+			readyFile=$readyDir/$identifier.html
+	
+			curl -s $ridingUrl > $ridingFile && mv -v $ridingFile $readyFile || true
 
-	for identifier in `cat $ridingList`; do
-		ridingUrl="http://enr.elections.ca/ElectoralDistricts.aspx?ed=$identifier&lang=$lang"
-		ridingFile=$workDir/$identifier.html
-		readFile=$readyDir/$identifier.html
-
-		curl -s $ridingUrl > $ridingFile && mv -v $ridingFile $readyFile || true
-
-		sleep 1 
-	done	
-
+			echo; echo 
+			echo Data for riding $identifier in language $lang ready!
+			echo; echo 
+	
+			sleep $pauseSecs
+		done	
+	
+	done
+	
+	date 
+	echo Resting for $restSecs seconds....
+	echo
+	echo To interupt until next cron:
+	echo rm $continueFlag
+	echo
+	echo To stop permantly
+	echo touch $stopFlag
+	echo
+	for i in `seq $restSecs -1 0`; do 
+		printf "$i... " ; 
+		if (( $i % 5 == 0 )) ; then echo ; fi
+		sleep 1; 
+	done
+	echo ; echo
 done
 
 
+echo; echo; echo; echo;
